+++ project/PlanetGenerator.h
@@ -2,12 +2,12 @@
 #include <iostream>
 #include <mutex>
 
-void generateEarth(unsigned int lengthX, unsigned int lengthY, std::string fileLocation, char* buffer, double* elevation,
-				   double* moisture, double* climate, bool& inProgress, int& progress, unsigned int seed = 0, std::mutex* phone = nullptr,
+void generateEarth(unsigned int lengthX, unsigned int lengthY, std::string fileLocation, char* buffer, float* elevation,
+				   float* moisture, float* climate, bool& inProgress, int& progress, unsigned int seed = 0, std::mutex* phone = nullptr,
 				   bool useMutex = false);
 
 double normalFunction(double x, double mean = 0.0, double deviation = 1.0);
-double normalizeToRange(double ix1, double ix2, double tx1, double tx2, double i);
+float normalizeToRange(float ix1, float ix2, float tx1, float tx2, float i);
 void returnRGBA(int val, int* rgb); //Takes in and modifies a size 4 int array representing a color
 
 int genDesert(int r, int c);
